@@ -1,14 +1,14 @@
 import React, { Fragment, useState, useEffect } from "react";
 import "./ForgotPassword.css";
 import Loader from "../layout/Loader";
-import MailOutlineIcon from "@material-ui/icons/MailOutline";
+import MailOutlineIcon from "@mui/icons-material/MailOutline";
 import { useDispatch, useSelector } from "react-redux";
 import { clearErrors, forgotPassword } from "../../actions/userAction";
-import { useAlert } from "react-alert";
+import { useSnackbar } from "notistack";
 import MetaData from "../layout/MetaData";
 const ForgotPassword = () => {
   const dispatch = useDispatch();
-  const alert = useAlert();
+  const { enqueueSnackbar } = useSnackbar();
 
   const { error, message, loading } = useSelector((state) => state.forgotPassword);
 
@@ -19,55 +19,55 @@ const ForgotPassword = () => {
     const myForm = new FormData();
     myForm.set("email", email);
     dispatch(forgotPassword(myForm));
-  }; 
+  };
   useEffect(() => {
     if (error) {
-      alert.error(error);
+      enqueueSnackbar(error, { variant: "error" });
       dispatch(clearErrors());
     }
 
     if (message) {
-    alert.success(message);
+      enqueueSnackbar(message, { variant: "success" });
     }
-  }, [dispatch, error, alert,message]);
-    return (
+  }, [dispatch, error, enqueueSnackbar, message]);
+  return (
+    <Fragment>
+      {loading ? (
+        <Loader />
+      ) : (
         <Fragment>
-        {loading ? (
-          <Loader />
-        ) : (
-          <Fragment>
-            <MetaData title="Forgot Password" />
-            <div className="forgotPasswordContainer">
-              <div className="forgotPasswordBox">
-                <h2 className="forgotPasswordHeading">Forgot Password</h2>
-  
-                <form
-                  className="forgotPasswordForm"
-                  onSubmit={forgotPasswordSubmit}
-                >
-                  <div className="forgotPasswordEmail">
-                    <MailOutlineIcon />
-                    <input
-                      type="email"
-                      placeholder="Email"
-                      required
-                      name="email"
-                      value={email}
-                      onChange={(e) => setEmail(e.target.value)}
-                    />
-                  </div>
+          <MetaData title="Forgot Password" />
+          <div className="forgotPasswordContainer">
+            <div className="forgotPasswordBox">
+              <h2 className="forgotPasswordHeading">Forgot Password</h2>
+
+              <form
+                className="forgotPasswordForm"
+                onSubmit={forgotPasswordSubmit}
+              >
+                <div className="forgotPasswordEmail">
+                  <MailOutlineIcon />
                   <input
-                    type="submit"
-                    value="Send"
-                    className="forgotPasswordBtn"
+                    type="email"
+                    placeholder="Email"
+                    required
+                    name="email"
+                    value={email}
+                    onChange={(e) => setEmail(e.target.value)}
                   />
-                </form>
-              </div>
+                </div>
+                <input
+                  type="submit"
+                  value="Send"
+                  className="forgotPasswordBtn"
+                />
+              </form>
             </div>
-          </Fragment>
-        )}
-      </Fragment>
-    )
+          </div>
+        </Fragment>
+      )}
+    </Fragment>
+  )
 }
 
 export default ForgotPassword
