@@ -1,82 +1,85 @@
-import React, { Fragment,useRef,useState,useEffect } from 'react'
+import React, { Fragment, useRef, useState, useEffect } from 'react'
 import "./LoginSignup.css"
 import Loader from "../layout/Loader"
-import { Link } from 'react-router-dom'
-import MailOutlineIcon from "@material-ui/icons/MailOutline"
-import LockOpenIcon from  "@material-ui/icons/LockOpen"
-import FaceIcon from "@material-ui/icons/Face"
-import {useDispatch,useSelector} from "react-redux"
-import {clearErrors,login,register} from "../../actions/userAction"
-import {useAlert} from "react-alert"
-const LoginSignup = ({history,location}) => {
+import { Link, useNavigate, useLocation } from 'react-router-dom'
+import MailOutlineIcon from "@mui/icons-material/MailOutline"
+import LockOpenIcon from "@mui/icons-material/LockOpen"
+import FaceIcon from "@mui/icons-material/Face"
+import { useDispatch, useSelector } from "react-redux"
+import { clearErrors, login, register } from "../../actions/userAction"
+import { useSnackbar } from "notistack"
+
+const LoginSignup = () => {
     const dispatch = useDispatch();
-    const alert = useAlert()
+    const { enqueueSnackbar } = useSnackbar();
+    const navigate = useNavigate();
+    const location = useLocation();
 
-    const loginTab= useRef(null)
-    const registerTab= useRef(null)
-    const switcherTab= useRef(null)
+    const loginTab = useRef(null)
+    const registerTab = useRef(null)
+    const switcherTab = useRef(null)
 
-    const {error,loading,isAuthenticated} = useSelector(state=>state.user)
-    const [loginEmail,setLoginEmail] = useState("")
-    const [loginPassword,setLoginPassword] = useState("")
-    const [user,setUser] = useState({
+    const { error, loading, isAuthenticated } = useSelector(state => state.user)
+    const [loginEmail, setLoginEmail] = useState("")
+    const [loginPassword, setLoginPassword] = useState("")
+    const [user, setUser] = useState({
         name: "",
         email: "",
         password: "",
     });
-    const [avatar,setAvatar] = useState("")
-    const [avatarPreview,setAvatarPreview] = useState("/Profile.png")
-    const {name,email,password} = user;
+    const [avatar, setAvatar] = useState("")
+    const [avatarPreview, setAvatarPreview] = useState("/Profile.png")
+    const { name, email, password } = user;
 
-    const loginSubmit = (e)=>{
+    const loginSubmit = (e) => {
         e.preventDefault();
-        dispatch(login(loginEmail,loginPassword))
+        dispatch(login(loginEmail, loginPassword))
     }
-    const registerSubmit = (e)=>{
+    const registerSubmit = (e) => {
         e.preventDefault();
         const myForm = new FormData()
-        myForm.set("name",name)
-        myForm.set("email",email)
-        myForm.set("password",password)
-        myForm.set("avatar",avatar)
+        myForm.set("name", name)
+        myForm.set("email", email)
+        myForm.set("password", password)
+        myForm.set("avatar", avatar)
         dispatch(register(myForm))
     }
-    const registerDataChange = (e) =>{
-        if (e.target.name==="avatar") {
+    const registerDataChange = (e) => {
+        if (e.target.name === "avatar") {
             const reader = new FileReader();
-            
-            reader.onload= ()=>{
-                if (reader.readyState===2) {
+
+            reader.onload = () => {
+                if (reader.readyState === 2) {
                     setAvatarPreview(reader.result)
                     setAvatar(reader.result)
                 }
             }
             reader.readAsDataURL(e.target.files[0]);
-        }else{
-            setUser({...user,[e.target.name]:e.target.value});
+        } else {
+            setUser({ ...user, [e.target.name]: e.target.value });
         }
     }
     const redirect = location.search ? location.search.split("=")[1] : "/account";
     useEffect(() => {
         if (error) {
-            alert.error(error)
+            enqueueSnackbar(error, { variant: "error" })
             dispatch(clearErrors)
         }
         if (isAuthenticated) {
-            history.push(redirect)
+            navigate(redirect)
         }
-    }, [dispatch,error,alert,history,isAuthenticated,redirect])
+    }, [dispatch, error, enqueueSnackbar, navigate, isAuthenticated, redirect])
 
 
-    const switchTabs = (e,tab) =>{
-        if (tab==="login") {
+    const switchTabs = (e, tab) => {
+        if (tab === "login") {
             switcherTab.current.classList.add("shiftToNeutral")
             switcherTab.current.classList.remove("shiftToRight")
 
             registerTab.current.classList.remove("shiftToNeutralForm")
             loginTab.current.classList.remove("shiftToLeft")
         }
-        if (tab==="register") {
+        if (tab === "register") {
             switcherTab.current.classList.remove("shiftToNeutral")
             switcherTab.current.classList.add("shiftToRight")
 
@@ -86,93 +89,93 @@ const LoginSignup = ({history,location}) => {
     }
     return (
         <Fragment>
-            {loading?(<Loader/>
-            ):(
+            {loading ? (<Loader />
+            ) : (
                 <Fragment>
-                <div className='LoginSignUpContainer'>
-                    <div className='LoginSignUpBox'>
-                        <div>
-                            <div className='login_signUp_toggle'>
-                                <p onClick={(e)=> switchTabs(e,"login")}>Login</p>
-                                <p onClick={(e)=> switchTabs(e,"register")}>Register</p>
+                    <div className='LoginSignUpContainer'>
+                        <div className='LoginSignUpBox'>
+                            <div>
+                                <div className='login_signUp_toggle'>
+                                    <p onClick={(e) => switchTabs(e, "login")}>Login</p>
+                                    <p onClick={(e) => switchTabs(e, "register")}>Register</p>
+                                </div>
+                                <button ref={switcherTab}></button>
                             </div>
-                            <button ref={switcherTab}></button>
-                        </div>
-                        <form className='loginForm' ref={loginTab} onSubmit={loginSubmit}>
-                            <div className='loginEmail'>
-                                <MailOutlineIcon/>
-                                <input 
-                                type="email"
-                                placeholder="Email"
-                                required
-                                value={loginEmail} 
-                                onChange={(e)=> setLoginEmail(e.target.value)}
-                                />
-                            </div>
-                            <div className='loginPassword'>
-                                <LockOpenIcon/>
-                                <input type="password"
-                                placeholder="Password"
-                                required
-                                value={loginPassword}
-                                onChange={(e)=> setLoginPassword(e.target.value)}
-                                />
-                            </div>
-                            <Link to="/password/forgot">Forgot Password ?</Link>
-                            <input type="submit" value="Login" className='loginBtn'/>
-                        </form>
-                        <form className='signUpForm' ref={registerTab} encType='multipart/form-data' onSubmit={registerSubmit}>
-                            <div className='signUpName'>
-                                <FaceIcon/>
-                                <input 
-                                type="text"
-                                placeholder="Name"
-                                required
-                                name= "name"
-                                value={name} 
-                                onChange={registerDataChange}
-                                />
-                            </div>
-                            <div className='signUpEmail'>
-                            <MailOutlineIcon/>
-                                <input 
-                                type="email"
-                                placeholder="Email"
-                                required
-                                name='email'
-                                value={email} 
-                                onChange={registerDataChange}
-                                />
-                            </div>
-                            <div className='signUpPassword'>
-                            <LockOpenIcon/>
-                                <input 
-                                type="password"
-                                placeholder="Password"
-                                required
-                                name='password'
-                                value={password}
-                                onChange={registerDataChange}
-                                />
-                            </div>
-                            <div id='registerImage'>
-                                <img src={avatarPreview} alt="avatar preview"/>
+                            <form className='loginForm' ref={loginTab} onSubmit={loginSubmit}>
+                                <div className='loginEmail'>
+                                    <MailOutlineIcon />
+                                    <input
+                                        type="email"
+                                        placeholder="Email"
+                                        required
+                                        value={loginEmail}
+                                        onChange={(e) => setLoginEmail(e.target.value)}
+                                    />
+                                </div>
+                                <div className='loginPassword'>
+                                    <LockOpenIcon />
+                                    <input type="password"
+                                        placeholder="Password"
+                                        required
+                                        value={loginPassword}
+                                        onChange={(e) => setLoginPassword(e.target.value)}
+                                    />
+                                </div>
+                                <Link to="/password/forgot">Forgot Password ?</Link>
+                                <input type="submit" value="Login" className='loginBtn' />
+                            </form>
+                            <form className='signUpForm' ref={registerTab} encType='multipart/form-data' onSubmit={registerSubmit}>
+                                <div className='signUpName'>
+                                    <FaceIcon />
+                                    <input
+                                        type="text"
+                                        placeholder="Name"
+                                        required
+                                        name="name"
+                                        value={name}
+                                        onChange={registerDataChange}
+                                    />
+                                </div>
+                                <div className='signUpEmail'>
+                                    <MailOutlineIcon />
+                                    <input
+                                        type="email"
+                                        placeholder="Email"
+                                        required
+                                        name='email'
+                                        value={email}
+                                        onChange={registerDataChange}
+                                    />
+                                </div>
+                                <div className='signUpPassword'>
+                                    <LockOpenIcon />
+                                    <input
+                                        type="password"
+                                        placeholder="Password"
+                                        required
+                                        name='password'
+                                        value={password}
+                                        onChange={registerDataChange}
+                                    />
+                                </div>
+                                <div id='registerImage'>
+                                    <img src={avatarPreview} alt="avatar preview" />
+                                    <input
+                                        type="file"
+                                        name='avatar'
+                                        accept='image/*'
+                                        onChange={registerDataChange}
+                                    />
+                                </div>
                                 <input
-                                type="file"
-                                name='avatar'
-                                accept='image/*'
-                                onChange={registerDataChange}
+                                    type="submit"
+                                    value="Register"
+                                    className='signUpBtn'
                                 />
-                            </div>
-                            <input
-                            type="submit"
-                            value="Register"
-                            className='signUpBtn'
-                            />
-                        </form>
+                            </form>
+                        </div>
                     </div>
-                </div>
-            </Fragment>
+                </Fragment>
             )}
         </Fragment>
     )
