@@ -3,16 +3,18 @@ import "./UpdatePassword.css";
 import Loader from "../layout/Loader";
 import { useDispatch, useSelector } from "react-redux";
 import { clearErrors, updatePassword } from "../../actions/userAction";
-import { useAlert } from "react-alert";
+import { useSnackbar } from "notistack";
 import { UPDATE_PASSWORD_RESET } from "../../constants/userConstants";
 import MetaData from "../layout/MetaData";
-import LockOpenIcon from  "@material-ui/icons/LockOpen"
-import LockIcon from  "@material-ui/icons/Lock"
-import VpnKeyIcon from "@material-ui/icons/VpnKey"
+import LockOpenIcon from "@mui/icons-material/LockOpen"
+import LockIcon from "@mui/icons-material/Lock"
+import VpnKeyIcon from "@mui/icons-material/VpnKey"
+import { useNavigate } from "react-router-dom";
 
-const UpdatePassword = ({history}) => {
+const UpdatePassword = () => {
   const dispatch = useDispatch();
-  const alert = useAlert();
+  const { enqueueSnackbar } = useSnackbar();
+  const navigate = useNavigate();
 
 
   const { error, isUpdated, loading } = useSelector((state) => state.profile);
@@ -29,23 +31,23 @@ const UpdatePassword = ({history}) => {
     myForm.set("newPassword", newPassword);
     myForm.set("confirmPassword", confirmPassword);
     dispatch(updatePassword(myForm));
-  }; 
+  };
 
 
-  useEffect(() => { 
+  useEffect(() => {
     if (error) {
-      alert.error(error);
+      enqueueSnackbar(error, { variant: "error" });
       dispatch(clearErrors());
     }
 
     if (isUpdated) {
-    history.push("/account");
-    alert.success("Profile Updated Successfully");
-    dispatch({
-    type: UPDATE_PASSWORD_RESET,
-    });
+      navigate("/account");
+      enqueueSnackbar("Profile Updated Successfully", { variant: "success" });
+      dispatch({
+        type: UPDATE_PASSWORD_RESET,
+      });
     }
-  }, [dispatch, error, alert, history,isUpdated]);
+  }, [dispatch, error, enqueueSnackbar, navigate, isUpdated]);
   return (
     <Fragment>
       {loading ? (
@@ -62,32 +64,32 @@ const UpdatePassword = ({history}) => {
                 encType="multipart/form-data"
                 onSubmit={updatePasswordSubmit}
               >
-                 <div className='loginPassword'>
-                                <VpnKeyIcon/>
-                                <input type="password"
-                                placeholder="Old Password"
-                                required
-                                value={oldPassword}
-                                onChange={(e)=> setOldPassword(e.target.value)}
-                                />
+                <div className='loginPassword'>
+                  <VpnKeyIcon />
+                  <input type="password"
+                    placeholder="Old Password"
+                    required
+                    value={oldPassword}
+                    onChange={(e) => setOldPassword(e.target.value)}
+                  />
                 </div>
                 <div className='loginPassword'>
-                                <LockOpenIcon/>
-                                <input type="password"
-                                placeholder="New Password"
-                                required
-                                value={newPassword}
-                                onChange={(e)=> setNewPassword(e.target.value)}
-                                />
+                  <LockOpenIcon />
+                  <input type="password"
+                    placeholder="New Password"
+                    required
+                    value={newPassword}
+                    onChange={(e) => setNewPassword(e.target.value)}
+                  />
                 </div>
                 <div className='loginPassword'>
-                                <LockIcon/>
-                                <input type="password"
-                                placeholder="Confirm Password"
-                                required
-                                value={confirmPassword}
-                                onChange={(e)=> setConfirmPassword(e.target.value)}
-                                />
+                  <LockIcon />
+                  <input type="password"
+                    placeholder="Confirm Password"
+                    required
+                    value={confirmPassword}
+                    onChange={(e) => setConfirmPassword(e.target.value)}
+                  />
                 </div>
                 <input
                   type="submit"
