@@ -1,10 +1,7 @@
 const mongoose = require('mongoose')
 // connnects to database 
 const connectDatabase = ()=>{
-    mongoose.connect(process.env.DB_URI, {
-        useNewUrlParser: true,
-        useUnifiedTopology: true,
-    }).then((data) => {
+    mongoose.connect(process.env.DB_URI).then((data) => {
         console.log(`mongodb connected with server: ${data.connection.host}`);
     })
 }
