@@ -1,6 +1,6 @@
 import React from 'react'
-import { Helmet } from 'react-helmet'
-const MetaData = ({title}) => {
+import { Helmet } from 'react-helmet-async'
+const MetaData = ({ title }) => {
     return (
         <Helmet>
             <title>{title}</title>
